# documentation/api.yaml
openapi: 3.0.0
info:
  title: Westco Tack API Course Project Documentation.
  version: 1.0
  licence:
    name: Apache 2.0
    url: "http://apache.org/"
servers:
  - url: http://localhost:5000
    description: Dev Server on Port
  # - url: http://staging-api.example.com
  #   description: Optional server description, e.g. Internal staging server for testing
components:
  securitySchemes:
    bearerAuth: 
      type: http
      scheme: bearer
      bearerFormat: JWT 
security:
  - bearerAuth: []
paths:
...
  /superAdmin/signUp:
    post:
      summary: Sign Up Super Admin
      tags:
        - Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                secretInfo:
                  type: string
              example:
                name: Sunna
                email: Sunna@gmail.com
                password: 12345Sunna
                secretInfo: Men Uzbekistonda yashayman!
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                tyep: object
                properties:
                _id: integer
                name: string
                email: string
                password: string
                secretInfo: string
                createdAt: string
                updatedAt: string
              example:
                _id: 62af3d203816e616bd006514
                name: Sunnatullox
                email: Sunnatullox@gmail.com
                password: $2a$10$vF7DFPtLrXo4FXpO2vm7U.GXPHpmJWgVWD51rVsXVgL84cAzjyPrW
                secretInfo: $2a$10$p/oRKgRopkRvx5/a/MTp/OEnhVk7wx9i8SIKJkVFPPipQGUUBSX2S
                createdAt: 2022-06-19T15:13:36.493+00:00
                updatedAt: 2022-06-20T14:01:43.788+00:00
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /superAdmin/signIn:
    post:
      summary: Sign In Super Admin
      tags:
        - Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
              example:
                email: Sunna@gmail.com
                password: 12345Sunna
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                tyep: object
                properties:
                token: integer
                _id: integer
                name: string
                email: string
                password: string
                secretInfo: string
                createdAt: string
                updatedAt: string
              example:
                token: TU3ODExNjB9.RoCiHr5wflJbYUWs67NIK4IL3PQcM6fTDg2eeBGsJWo
                _id: 62af3d203816e616bd006514
                name: Sunnatullox
                email: Sunnatullox@gmail.com
                password: $2a$10$vF7DFPtLrXo4FXpO2vm7U.GXPHpmJWgVWD51rVsXVgL84cAzjyPrW
                secretInfo: $2a$10$p/oRKgRopkRvx5/a/MTp/OEnhVk7wx9i8SIKJkVFPPipQGUUBSX2S
                createdAt: 2022-06-19T15:13:36.493+00:00
                updatedAt: 2022-06-20T14:01:43.788+00:00
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /superAdmin/getAllUsers:
    get:
      summary: Get All Users
      tags:
        - Super Admin
      description: Get All Users registered Platform
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                tyep: object
                properties:
                token: integer
                _id: integer
                name: string
                email: string
                password: string
                userLessons: array
                userAdmin: boolen
                userCourseId: array
              example:
                count: 1
                users:
                  token: TU3ODExNjB9.RoCiHr5wflJbYUWs67NIK4IL3PQcM6fTDg2eeBGsJWo
                  _id: 62ae948249ef0d92e405b22b
                  name: Sunnatullox
                  email: Sunnatullox@gmail.com
                  password: $2a$10$4JUm3qBt.FPGxnqcfYZ08./loDWvYA8VpHxDYdMqVBccieZv/vuBua
                  userLessons: []
                  userAdmin: false
                  userCourseId: []
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /superAdmin/getOneUser/{id}:
    get:
      summary: Get One User
      tags:
        - Super Admin
      description: Get one user registered platform
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Get User One for user Id
          example: 62ac7f3fa525215eeede092f
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                tyep: object
                properties:
                _id: integer
                name: string
                email: string
                password: string
                userLessons: array
                userAdmin: boolen
                userCourseId: array
              example:
                count: 1
                users:
                  _id: 62ae948249ef0d92e405b22b
                  name: Sunnatullox
                  email: Sunnatullox@gmail.com
                  password: $2a$10$4JUm3qBt.FPGxnqcfYZ08./loDWvYA8VpHxDYdMqVBccieZv/vuBua
                  userLessons: []
                  userAdmin: false
                  userCourseId: []
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /superAdmin/createAdmin/{id}:
    put:
      summary: Created user Admin
      tags:
        - Super Admin
      description: Get All Course Platform
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Get User One for user Id
          example: 62ac7f3fa525215eeede092f
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description: created admin required 'false' && 'true'
                userAdmin:
                  type: string
              example:
                userAdmin: "true"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                tyep: object
                properties:
                _id: integer
                name: string
                email: string
                password: string
                userLessons: array
                userAdmin: boolen
                userCourseId: array
              example:
                count: 1
                users:
                  _id: 62ae948249ef0d92e405b22b
                  name: Sunnatullox
                  email: sunnatullox@gmail.com
                  password: $2a$10$4JUm3qBt.FPGxnqcfYZ08./loDWvYA8VpHxDYdMqVBccieZv/vuBua
                  userLessons: []
                  userAdmin: false
                  userCourseId: []
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /superAdmin/updateUser/{id}:
    put:
      summary: Super Admin updated user
      tags:
        - Super Admin
      description: Get All Course Platform
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Get User One for user Id
          example: 62ac7f3fa525215eeede092f
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: string
                email: string
                password: string
              example:
                name: Sunnatullox
                email: sunnatullox@gmail.com
                password: 12345
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                tyep: object
                properties:
                _id: integer
                name: string
                email: string
                password: string
                userLessons: array
                userAdmin: boolen
                userCourseId: array
              example:
                count: 1
                users:
                  _id: 62ae948249ef0d92e405b22b
                  name: Sunnatullox
                  email: Sunnatullox@gmail.com
                  password: $2a$10$4JUm3qBt.FPGxnqcfYZ08./loDWvYA8VpHxDYdMqVBccieZv/vuBua
                  userLessons: []
                  userAdmin: false
                  userCourseId: []
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /superAdmin/deleteUser/{id}:
    delete:
      summary: Super Admin delete user
      tags:
        - Super Admin
      description: Get All Course Platform
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Get User One for user Id
          example: 62ac7f3fa525215eeede092f
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                user: string
              example:
                user: User Deleted
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /superAdmin/superAdminUpdatePass/acountDataCheck:
    post:
      summary: Forgot password? Super Admin check Password
      tags:
        - Super Admin
      description: When the Super Admin forgets the password, check the information to change the password
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email: string
                secretInfo: string
              example:
                email: Sunna@gmail.com
                secretInfo: Men Uzbekistonda yashayman!
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                tyep: object
                properties:
                Checked: boolen
              example:
                Checked: true
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /superAdmin/updatePassword:
    put:
      summary: Forgot password? Super Admin Update Password
      tags:
        - Super Admin
      description: When the Super Admin forgets the password, check the information to change the password
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password: string
                adminId: string
              example:
                password: "123456789"
                adminId: 62b17319c6d4c0fe6ff19aa1
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                tyep: object
                properties:
                password: string
              example:
                password: Secsess fully updated
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
  /superAdmin/updateProfile:
    post:
      summary: Super Admin Update Profile
      tags:
        - Super Admin
      description: To change the Super Admin cabinet data, you must first enter the old password and the security word and then the new ones
      consumes:
        - multipart/form-data
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                adminId:
                  type: string
                name:
                  type: string
                email:
                  type: string
                oldPassword:
                  type: string
                newPassword:
                  type: string
                oldSecretInfo:
                  type: string
                newSecretInfo:
                  type: string
                adminAvatar:
                  type: string
                  format: binary
      responses:
        200:
          description: OK
          content:
            multipart/form-data:
              schema:
                tyep: object
                properties:
                adminProfile: string
              example:
                adminProfile: Secsess fully updated Super Admin Profile
        404:
          description: Not Found Course
        500:
          description: Internal Server Error
...